x expression was implicitly casted from 'int' type to 'double' type
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@5614 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/Audio/File_Dts.cpp
@@ -1229,7 +1229,7 @@ float64 File_Dts::BitRate_Get(bool WithHD)
                 default     :   SamplePerFrames=    0; break; //Can never happen (4 bits)
             }
             if (SamplePerFrames)
-                BitRate+=HD_size*8*DTS_HD_MaximumSampleRate[HD_MaximumSampleRate]/SamplePerFrames;
+                BitRate+=((float64)HD_size)*8*DTS_HD_MaximumSampleRate[HD_MaximumSampleRate]/SamplePerFrames;
         }
         //if (Primary_Frame_Byte_Size && Profile==__T("HRA"))
         //    BitRate*=1+((float64)HD_size)/Primary_Frame_Byte_Size; //HD block are not in the nominal bitrate
